promotions remarks corrected/4
