+++ app/code/Study/Promotions/Api/PromotionsValidatorInterface.php
@@ -6,13 +6,6 @@ namespace Study\Promotions\Api;
 interface PromotionsValidatorInterface
 {
     /**
-     * Promotion status values
-     */
-    public const PROMOTION_ENABLED = 1;
-
-    public const PROMOTION_DISABLED = 0;
-
-    /**
      * Check is time parameters is correct
      *
      * @param string $startTime
