+++ app/code/Study/Promotions/Model/PromotionsValidator.php
@@ -11,13 +11,6 @@ use Study\Promotions\Model\PromotionsRepository;
  */
 class PromotionsValidator implements PromotionsValidatorInterface
 {
-    private $promotionsRepository;
-
-    public function __construct(
-        PromotionsRepository $promotionsRepository
-    ) {
-        $this->promotionsRepository = $promotionsRepository;
-    }
     /**
      * Check is time parameters is correct
      *
@@ -40,47 +33,4 @@ class PromotionsValidator implements PromotionsValidatorInterface
         return $result;
     }
 
-    /**
-     * Check is that promotion is enabled
-     *
-     * @param int $promotionId
-     * @return bool
-     */
-    public function isPromotionEnabled(int $promotionId): bool
-    {
-        if ($this->promotionsRepository->getById($promotionId)->getStatus() == self::PROMOTION_ENABLED) {
-
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Check is that promotion active in current moment of time
-     *
-     * @param int $promotionId
-     * @return bool
-     */
-    public function checkPromotionIsActiveNow(int $promotionId): bool
-    {
-        $promotion = $this->promotionsRepository->getById($promotionId);
-        $start = strtotime($promotion->getStartTime());
-        $finish = $promotion->getFinishTime();
-
-        if ($promotion->getFinishTime()) {
-            $finish = strtotime($promotion->getFinishTime());
-        }
-
-        if (!$finish && $start <= time()) {
-
-            return true;
-        }
-        if ($start <= time() && time() <= $finish) {
-
-            return true;
-        }
-
-        return false;
-    }
 }
